syntax: don't unilaterally clear syntax
From :help :syn-clear

------------------------------------------------------------------------
If you want to clear the syntax stuff for the current buffer, you can
use this command: >
  :syntax clear

This command should be used when you want to switch off syntax
highlighting, or when you want to switch to using another syntax.  It's
normally not needed in a syntax file itself, because syntax is cleared
by the autocommands that load the syntax file.
The command also deletes the "b:current_syntax" variable, since no
syntax is loaded after this command.
------------------------------------------------------------------------