indent: re-indent calc operator lines when typed
Many of the set indentkeys are redundant (vim takes the shortest prefix,
so that when you start <==, the < triggers re-indenting). But this makes
writing calcs a little nicer, as they automatically line up for you.

0= is necessary: =word triggers re-indent on the last character of
{word}, and 0 means it only happens when there is whitespace before
{word}. This is exactly the case with calc, though may also be true on
continuation lines as well. I currently don't have any real support for
those baked in, so this is a non-issue at the moment.